+++ src/app/app.component.ts
@@ -1,4 +1,8 @@
 import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
+import { CartService } from './cart/services/cart.service';
+import { Role } from './core/enums/role';
+import { AuthService } from './core/services/auth.service';
+import { SpinnerService } from './widgets';
 
 @Component({
   selector: 'app-root',
@@ -6,10 +10,24 @@ import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements AfterViewInit {
-  title = 'Angular Training - Task 2';
+  title = 'BoozeMart';
   @ViewChild('appTitle') titleElement: ElementRef<HTMLHeadingElement>;
+
+  get cartCount(): number {
+    return this.cartService.getTotalQuantity();
+  }
+
+  constructor(
+    private cartService: CartService,
+    private authService: AuthService,
+    public spinnerService: SpinnerService,
+  ) {}
 
   ngAfterViewInit(): void {
     this.titleElement.nativeElement.innerText = this.title;
+  }
+
+  isLoggedIn(): boolean {
+    return this.authService.isLoggedIn && this.authService.role === Role.Admin;
   }
 }