+++ src/app/order/models/order.model.ts
@@ -1,0 +1,7 @@
+import { CartItem } from '../../cart/models/cart-item.model';
+
+export class Order {
+  constructor(
+    public cartItems: Array<CartItem>
+  ) { }
+}