+++ src/app/products/products.module.ts
@@ -2,17 +2,21 @@ import { NgModule } from '@angular/core';
 import { ProductComponent } from './components/product/product.component';
 import { ProductListComponent } from './components/product-list/product-list.component';
 import { SharedModule } from '../shared/shared.module';
+import { ProductViewComponent } from './components/product-view/product-view.component';
+import { ProductsRoutingModule } from './products-routing.module';
 
 @NgModule({
   declarations: [
     ProductComponent,
-    ProductListComponent
+    ProductListComponent,
+    ProductViewComponent,
   ],
   imports: [
-    SharedModule
+    SharedModule,
+    ProductsRoutingModule
   ],
   exports: [
-    ProductListComponent
+    ProductComponent
   ]
 })
 export class ProductsModule { }